core_calendar: add indices to the events table
Add some indices to the "events" table to improve query performance in
raw_event_retrieval_strategy::get_raw_events_legacy_implementation().

This will likely only affect MySQL / MariaDB installations.

It also moves the `WHERE` statement from the `$unionstartquery` to
the individual subqueries in the `$subqueryconditions` array. This
change should preserve the query (modulo whitespace changes), since the
`$unionstartquery` variable is prepended as well as used as a joiner
when imploding the `$subqueryconditions`.

On mysql-based db systems, the queries built by `raw_event_retrieval_strategy`
are supplemented by adding a `FORCE INDEX` statements between the `FROM`
clause and the `WHERE` clause.

This significantly improves performance on instances with a large number of
rows in the events table (we observed improvements from ~4m to ~500ms).

This patch has not been tested on PostgreSQL systems.

Co-authored-by: Stefan Auditor <stefan@auditor.email>
Co-authored-by: Fabian Freyer <fabian.freyer@physik.tu-berlin.de>